requirements.txt
Django~=5.1.4
djangorestframework
PyYAML
drf-yasg